Modido tipo de dado de duas colunas do banco Removido pasta "migrations" do gitignore

+++ application/barganhou/register/models.py
@@ -3,8 +3,8 @@ from django.db import models
 # Create your models here.
 class ProductInfo (models.Model):
     name = models.CharField(max_length=20)
-    price = models.FloatField()
-    amount = models.DecimalField(max_digits=10, decimal_places=2)
+    price = models.DecimalField(max_digits=11, decimal_places=2)
+    amount = models.IntegerField()
     date_log = models.DateField()
     local = models.CharField(max_length=30)
     pucharse = models.CharField(max_length=3,
